Guided Tours: improve React context provided by the `makeTour` wrapper
Several cleanups:

1. `getChildContext` doesn't need to merge the parent context (`this.context`) and
the provided context (`this.tourMeta`). This was added in #22041 to make Redux `store`
available in the tour context, but it's not needed. React does that for us, as it
supports multiple nested context providers. All we need to do is to define the right
`childContextTypes` (without `store`, so that it's never overwritten) and `contextTypes`
(with `store`, so that it's not masked out and is part of the consumer's context).

I also refactored the `makeTour` component to use `getDerivedStateFromProps` to compute
new context every time its props change and to use `this.state` as the source of the
context. This will make migration to the new React Context API easier.