SDK: Eat Notifications Panel - Take 2
In a previous attempt I tried to create a webpack config inside of the
notifications directory but in this attempt I'm using learnings from the
SDK project and creating a new SDK script runner in order to build the
code in the subdirectory from the parent commands.

This builds a bloated image which is currently failing. I think
`core-js` is somehow getting in and causing problems.

Build with this command:
```
npm run sdk:notifications --output-dir=~/whatever
```

Then go to `~/whatever` and run a local webserver. The notifications
panel should load at the address (use `notifications.dev:3000`)